11/27

+++ src/main/java/Repository/LearningListRepository.java
@@ -7,5 +7,5 @@ import entity.LearningReport;
 
 @Repository
 public interface LearningListRepository
-    extends JpaRepository <LearningReport, Long> {
+    extends JpaRepository <LearningReport, Integer> {
 }
